+++ nixos/displayManager.nix
@@ -16,9 +16,9 @@
 			xkb.variant = "";
 		};
 		displayManager = {
-			sddm.enable = false;
+			sddm.enable = true;
 		};
-		desktopManager.plasma6.enable = false;
+		desktopManager.plasma6.enable = true;
 		};
 
 	# Esto habilita Hyprland.